Random answer shouldn't be annoying to guess
Closes https://github.com/iancanderson/nerdle/issues/8
